+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "server",
+  "version": "1.1.0",
+  "license": "MIT",
+  "author": "HackMerced",
+  "main": "server.js",
+  "scripts": {
+    "test": "mocha",
+    "clean": "concurrently \"yarn clean:client\" \"yarn clean:server\"",
+    "clean:client": "cd client && rm -rf ./node_modules yarn.lock",
+    "clean:server": "rm -rf ./node_modules yarn.lock",
+
+    "dev": "concurrently \"yarn dev:client\" \"yarn dev:server\" && echo \"HackMerced V up and running\"",
+    "dev:client": "cd ./client && yarn build",
+    "dev:server": "nodemon server.js"
+  },
+  "dependencies": {
+    "body-parser": "^1.19.0",
+    "concurrently": "^4.1.2",
+    "dotenv": "^8.1.0",
+    "express": "^4.17.1",
+    "mongodb": "^3.2.7",
+    "morgan": "^1.9.1"
+  },
+  "devDependencies": {
+    "mocha": "^6.1.4",
+    "nodemon": "^1.19.1"
+  }
+}
