--- server/.env.tmpl
@@ -1,1 +1,0 @@
-uri=mongodb+srv://<YOUR-USERNAME>:<YOUR-PASSWORD>@<YOUR-CLUSTER>.mongodb.net/<YOUR-COLLECTION>?retryWrites=true&w=majority
