--- server/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "license": "MIT",
-  "author": "HackMerced",
-  "main": "server.js",
-  "scripts": {
-    "test": "mocha",
-    "start": "nodemon server.js"
-  },
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "dotenv": "^8.0.0",
-    "express": "^4.17.1",
-    "mongodb": "^3.2.7",
-    "morgan": "^1.9.1"
-  },
-  "devDependencies": {
-    "mocha": "^6.1.4",
-    "nodemon": "^1.19.1"
-  }
-}
